+++ fin-api/cli/bin/fin-io.js
@@ -6,6 +6,7 @@ const program = new Command();
 program
   .command('import <root-fs-path>')
   .option('-m, --force-metadata-update', 'Always re-PUT metadata, ignore sha check')
+  .option('-b, --force-binary-update', 'Always re-PUT binary, ignore sha check')
   .option('-r, --dry-run', 'do not write any containers')
   .option('--import-from-root', 'import data starting at given fs path, instead of ArchivalGroups')
   .option('--fcrepo-path-type <type>', 'import relative to ArchivalGroups + subpath + id (id, default) or subpath + id (subpath)')